Fix the model wallet and BigDecimal values

+++ backend/src/main/java/com/mmsl/fiwmoney/dto/StockDTO.java
@@ -3,14 +3,16 @@ package com.mmsl.fiwmoney.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class StockDTO {
 
     private Long id;
     private String code;
-    private double currentPrice;
-    private double averagePrice;
+    private BigDecimal currentPrice;
+    private BigDecimal averagePrice;
     private boolean notify;
 
 }
